+++ .dockerignore
@@ -1,3 +1,16 @@
-react/node_modules
-.git
+# Developer configs
+.editorconfig
 .gitignore
+
+# Project utilities
+biome.json
+lefthook.yml
+*.env
+
+# Project folders
+.github
+docker
+node_modules
+
+# Documentation
+*.md
