+++ .github/dependabot.yml
@@ -1,7 +1,7 @@
 version: 2
 updates:
   - package-ecosystem: "npm"
-    directory: "/react/"
+    directory: "/"
     schedule:
       interval: "weekly"
     commit-message:
