+++ .github/workflows/lint.yml
@@ -25,6 +25,6 @@ jobs:
         uses: biomejs/setup-biome@v2
 
       - name: Run Biome
-        run: biome ci --error-on-warnings --formatter-enabled=false --organize-imports-enabled=false react/src
+        run: biome ci --error-on-warnings --formatter-enabled=false --organize-imports-enabled=false src
