--- react/package.json
@@ -1,39 +1,0 @@
-{
-	"name": "ucmacm-website",
-	"private": true,
-	"version": "0.0.0",
-	"type": "module",
-	"scripts": {
-		"dev": "vite",
-		"build": "tsc && vite build",
-		"preview": "vite preview",
-		"check": "biome check --write src",
-		"prepare": "lefthook install",
-		"serve": "serve /website/dist --no-clipboard"
-	},
-	"dependencies": {
-		"@icons-pack/react-simple-icons": "^10.2.0",
-		"@nextui-org/react": "^2.6.8",
-		"framer-motion": "^11.14.4",
-		"react": "^18.3.1",
-		"react-dom": "^18.3.1",
-		"react-router-dom": "^7.1.0",
-		"supertokens-auth-react": "^0.48.0"
-	},
-	"devDependencies": {
-		"@biomejs/biome": "1.9.4",
-		"@types/react": "^18.3.1",
-		"@types/react-dom": "^18.3.1",
-		"@vitejs/plugin-react-swc": "^3.7.2",
-		"autoprefixer": "^10.4.20",
-		"lefthook": "^1.10.0",
-		"postcss": "^8.4.49",
-		"serve": "^14.2.4",
-		"tailwindcss": "^3.4.16",
-		"typescript": "^5.7.2",
-		"vite": "^6.0.3"
-	},
-	"volta": {
-		"node": "22.12.0"
-	}
-}
