--- react/postcss.config.js
@@ -1,6 +1,0 @@
-export default {
-	plugins: {
-		tailwindcss: {},
-		autoprefixer: {},
-	},
-};
