--- react/vite.config.ts
@@ -1,7 +1,0 @@
-import react from "@vitejs/plugin-react-swc";
-import { defineConfig } from "vite";
-
-// https://vitejs.dev/config/
-export default defineConfig({
-	plugins: [react()],
-});
